공통코드 interface 모듈화
