+++ src/models/interface/system/commoncode/IDetailCode.tsx
@@ -1,0 +1,6 @@
+export interface IDetailCode {
+    groupCode: string,
+    detailCode: string,
+    detailName: string,
+    note: string,
+}
