+++ src/models/interface/system/commoncode/IDetailCodeDetail.tsx
@@ -1,0 +1,6 @@
+import { IDetailCode } from "./IDetailCode";
+
+export interface IDetailCodeDetail extends IDetailCode {
+    useYn: "Y" | "N",
+    higherCode: string
+}
