+++ src/models/interface/system/commoncode/IGroupCode.tsx
@@ -1,0 +1,6 @@
+export interface IGroupCode {
+    groupCode: string,
+    groupName: string,
+    note: string,
+    useYn: "Y" | "N",
+}
